+++ crates/webui-handler/src/plugin/fast_v3.rs
@@ -164,7 +164,7 @@ impl HandlerPlugin for FastV3HydrationPlugin {
         state: &Value,
         writer: &mut dyn ResponseWriter,
     ) -> Result<()> {
-        super::fast_route_state::write_fast_route_component_state(state, writer)
+        super::fast::write_route_component_state(state, writer)
     }
 }
