+++ docs/ai/SKILL.md
@@ -911,10 +911,7 @@ identifier is falsy, so `path` is false and `!path` is true. No error.
 
 **Route-scoped state.** Each route handler should return only the keys that
 route's template binds to. Sending full app state on every route wastes
-bandwidth and render time. FAST route components expose scalar values as
-kebab-case HTML attributes, so keep `@attr` inputs flat. Strings, numbers, and
-booleans are emitted. Complex values should be initialized from the rendered
-DOM or another documented mechanism appropriate to the component.
+bandwidth and render time.
 
 ### Reserved `$webui` state
